cli_ml_automation/requirements.txt
requests==2.18.4
tensorflow==1.5.0
Flask==0.12.2
pandas==0.23.0
scikit_learn==0.19.1
sklearn_pandas==1.6.0
catboost==0.7
h5py==2.6.0
keras==2.1.5
filelock==3.0.4
pytest==3.0.5
xgboost==0.80
numpy==1.14.5
lightgbm==2.1.2
boto3==1.9.171
common_utils==0.0.1